+++ src/charCode.ts
@@ -1,38 +1,76 @@
 export const enum CharCode {
     /**
+     * The NUL character.
+     */
+    Null = 0,
+    /**
      * The `\t` character.
      */
     Tab = 9,
     /**
+     * The Unit Separator character.
+     */
+    UnitSeparator = 31,
+    /**
      * The ` ` character.
      */
     Space = 32,
+
     /**
      * The `!` character.
      */
     Exclamation = 33,
     /**
-	 * The `/` character.
-	 */
+     * The `/` character.
+     */
     Slash = 47,
-	/**
-	 * The `:` character.
-	 */
+
+    Zero = 48,
+    Nine = 57,
+
+    /**
+     * The `:` character.
+     */
     Colon = 58,
+
+    /**
+     * The `@` character.
+     */
+    At = 64,
 
     A = 65,
     Z = 90,
+
+    /**
+     * The `[` character.
+     */
+    LeftBracket = 91,
 
     /**
      * The `\` character.
      */
     Backslash = 92,
 
+    /**
+     * The ` character.
+     */
+    Backtick = 96,
+
     a = 97,
     z = 122,
 
     /**
+     * The `{` character.
+     */
+    LeftBrace = 123,
+
+    /**
      * The `~` character.
      */
-    Tide = 126
+    Tide = 126,
+
+    /**
+     * The DEL character.
+     */
+    Delete = 127
 }